some changes

+++ Main/Contracts/JobContracts/ContractJobTableViewCell.swift
@@ -31,12 +31,12 @@ class ContractJobTableViewCell: UITableViewCell {
     func setViews(contract: GetContractsJobItem) {
         nameLbl.text = contract.name
         createdAtLbl.text = contract.created_at?.toDate()?.timeAgoDisplay()
-        if let hours = contract.hours?.removeZerosFromEnd() {
+        if let hours = contract.hours?.toDouble.removeZerosFromEnd() {
             let mutableString = NSMutableAttributedString(string: "\(hours) h", attributes: [NSAttributedString.Key.font: UIFont.boldSystemFont(ofSize: 20)])
             mutableString.addAttribute(NSAttributedString.Key.foregroundColor, value: UIColor.systemGray2, range: NSRange(location: hours.count + 1, length: 1))
             hoursLbl.attributedText = mutableString
         }
-        if let balance = contract.amount?.removeZerosFromEnd() {
+        if let balance = contract.amount?.toDouble.removeZerosFromEnd() {
             let mutableString = NSMutableAttributedString(string: "\(balance) sc", attributes: [NSAttributedString.Key.font: UIFont.boldSystemFont(ofSize: 20)])
             mutableString.addAttribute(NSAttributedString.Key.foregroundColor, value: UIColor.systemGray2, range: NSRange(location: balance.count + 1, length: 2))
             balanceLbl.attributedText = mutableString
